+++ app/Models/TeamStatus.php
@@ -5,14 +5,14 @@ namespace App\Models;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 
-class CreatorStatus extends Model
+class TeamStatus extends Model
 {
     use HasFactory;
 
-    protected $table = 'creator_statuses';
+    protected $fillable = ['status'];
 
-    public function creator()
+    public function teams()
     {
-        return $this->hasMany(Creator::class);
+        return $this->hasMany(Team::class, 'team_status_id');
     }
 }