+++ database/migrations/2024_01_10_085055_rename_name_to_status_in_creator_statuses_table.php
@@ -1,0 +1,22 @@
+<?php
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+
+class RenameNameToStatusInCreatorStatusesTable extends Migration
+{
+    public function up()
+    {
+        Schema::table('creator_statuses', function (Blueprint $table) {
+            $table->renameColumn('name', 'status');
+        });
+    }
+
+    public function down()
+    {
+        Schema::table('creator_statuses', function (Blueprint $table) {
+            $table->renameColumn('status', 'name');
+        });
+    }
+};